Actually factor out different behaviors and resolve
