+++ pixi.toml
@@ -313,8 +313,15 @@ misc-fmt-check = "prettier --check '**/*.{yml,yaml,js,css,html}'"
 # Run first ruff fix, then ruff format, order is important see also https://twitter.com/charliermarsh/status/1717229721954799727
 py-fmt = "ruff check --fix --config rerun_py/pyproject.toml . && ruff format --config rerun_py/pyproject.toml ."
 py-fmt-check = "ruff check --config rerun_py/pyproject.toml . && ruff format --check --config rerun_py/pyproject.toml"
-py-lint = "mypy --config-file rerun_py/pyproject.toml --install-types --non-interactive --no-warn-unused-ignore"
-
+# Parametrize py-lint
+py-lint-config = {args=["config", "args"], cmd = "mypy --config-file {{config}} --install-types --non-interactive --no-warn-unused-ignore {{args}}" }
+# Get non-internal things that are ok to not check rerun on
+py-lint-non-sdk = { depends-on = [{"task" = "py-lint-config", "args" = ["rerun_py/.non_sdk_mypy.ini", ""] }] }
+# Need to run this in env with rerun installed
+py-lint-rerun = { depends-on = [{"task" = "py-lint-config", "args" = ["rerun_py/pyproject.toml", ""] }] }
+# Backwards compatible version
+py-lint-rerun-old = { depends-on = [{"task" = "py-lint-config", "args" = ["rerun_py/pyproject.toml", "--disable-error-code=import-not-found"] }] }
+py-lint = {depends-on = ["py-lint-non-sdk", "py-lint-rerun-old"] }
 
 rs-plot-dashboard = { cmd = "cargo r -p plot_dashboard_stress --release --" }
 
@@ -596,7 +603,7 @@ pytest = ">=7"
 pytest-benchmark = ">=4.0.0,<4.1"
 python = "=3.11"                  # We use the latest Python version here, so we get the latest mypy etc, EXCEPT 3.12 is too new for some of our examples. We run our CI tests on ALL supported versions though.
 ruff = "0.9.6.*"
-semver = ">=2.13,<2.14"
+semver = ">=3.0,<3.1"
 taplo = "=0.9.1"
 tomlkit = "0.12.3.*"
 tqdm = ">=4.66.2,<4.67"           # For displaying progress in various utility scripts.
