+++ rerun_py/.non_sdk_mypy.ini
@@ -1,0 +1,25 @@
+[mypy]
+files = examples/python, docs/snippets, scripts
+# hopeless examples
+exclude = (?x)(
+    ^examples/python/objectron/.*
+    | ^examples/python/ros_node/.*
+    | docs/snippets/all/concepts/how_helix_was_logged.py
+    | docs/snippets/all/concepts/static/log_static.py
+    | docs/snippets/all/concepts/static/log_static_10x.py
+    | docs/snippets/all/concepts/static/log_temporal_10x.py
+    | docs/snippets/all/concepts/static/send_static.py
+    | docs/snippets/all/concepts/static/send_static_10x.py
+    | ^docs/snippets/all/migration/log_line/.*
+    | docs/snippets/all/tutorials/custom-application-id.py
+    | docs/snippets/all/tutorials/custom-recording-id.py
+    | docs/snippets/all/tutorials/timelines_example.py
+    )
+namespace_packages = true
+show_error_codes = true
+strict = true
+enable_error_code = redundant-expr, truthy-bool, ignore-without-code
+plugins = numpy.typing.mypy_plugin
+ignore_missing_imports = true
+no_implicit_reexport = false
+disallow_untyped_calls = false
