+++ rerun_py/pyproject.toml
@@ -19,7 +19,8 @@ dependencies = [
   "numpy>=2",
   "pillow>=8.0.0",          # Used for JPEG encoding. 8.0.0 added the `format` arguments to `Image.open`
   "pyarrow>=18.0.0",
-  "typing_extensions>=4.5", # Used for PEP-702 deprecated decorator
+  "typing_extensions>=4.5",
+  "rerun_notebook" # I'm not sure if we really need to pin this at all today
 ]
 description = "The Rerun Logging SDK"
 keywords = ["computer-vision", "logging", "rerun"]
@@ -210,32 +211,14 @@ norecursedirs = ".* venv* target* build"
 ######################
 
 [tool.mypy]
+# Examples and docs are managed independently of the SDK.
+# in ini file adjacent to this.
 files = [
-  "docs/snippets",
   "rerun_py/rerun_sdk/rerun",
   "rerun_py/rerun_bindings",
   "rerun_py/tests",
   "rerun_notebook/src/rerun_notebook",
-  "examples/python",
-  "scripts",
   "tests/python",
-]
-exclude = [
-  # snippets that are not runnable
-  "docs/snippets/all/concepts/how_helix_was_logged",
-  "docs/snippets/all/concepts/static/log_static",
-  "docs/snippets/all/concepts/static/log_static_10x",
-  "docs/snippets/all/concepts/static/log_temporal_10x",
-  "docs/snippets/all/concepts/static/send_static",
-  "docs/snippets/all/concepts/static/send_static_10x",
-  "docs/snippets/all/migration/log_line",
-  "docs/snippets/all/tutorials/custom-application-id",
-  "docs/snippets/all/tutorials/custom-recording-id",
-  "docs/snippets/all/tutorials/timelines_example",
-
-  # hopeless examples
-  "examples/python/objectron",
-  "examples/python/ros_node",
 ]
 namespace_packages = true
 show_error_codes = true
@@ -247,5 +230,14 @@ no_implicit_reexport = false
 disallow_untyped_calls = false
 
 [[tool.mypy.overrides]]
-module = ["pyarrow.*"]
+
+module = [
+  "pyarrow.*",
+  # TODO[nick]: remove datafusion after #10696 lands
+  "datafusion.*",
+  # https://github.com/jupyter-widgets/ipywidgets/issues/3688
+  "ipywidgets.*",
+  # TODO[nick]: is this blocked on above?
+  "jupyter_ui_poll.*"
+]
 ignore_missing_imports = true
