+++ rerun_py/rerun_sdk/rerun/datatypes/bool.py
@@ -41,7 +41,7 @@ else:
     BoolLike = Any
 
 BoolArrayLike = Union[
-    BoolLike,
+    Bool,
     Sequence[BoolLike],
 ]
